Hehe

+++ pidController.c
@@ -23,7 +23,7 @@
 
 #define ALT_KP              50         // Altitude proportional gain
 #define ALT_KI              20          // Altitude integral gain
-#define ALT_KD              20         // Altitude derivative gain // Bro why does this have to be negative?
+#define ALT_KD              20         // Altitude derivative gain
 #define YAW_KP              100         // Yaw proportional gain
 #define YAW_KI              30//20          // Yaw integral gain
 #define YAW_KD              30//20          // Yaw derivative gain
@@ -96,7 +96,7 @@ getControlSignal(controller_t* piController, int32_t reference, int32_t measurem
     int32_t dutyCycle;
     int32_t controlSignal;
 
-    float errorSignal;
+    int32_t errorSignal;
     float derivativeError;
